move SimulationSharedInputs to simulation submodule
